Pass parent ref when dereferencing nested subschemas
When dereferencing a subschema that we got to via a
global reference, we might end up in a situation where
we have a subschema that has no reference, but
does have children.

We recurse to dereference the children, but if those
children in turn have references to definitions that
are local to this schema, the resolve_pointer method
has no way of figuring out which schema it is in, and
will default to looking it up on @schema... which is
not the correct one.

It won't find anything, and will complain about
unresolved pointers.

This passes along a parent reference when dereferencing
child schemas, so that it can use it to reconstruct the
correct reference if necessary.